Moved function out of utils so Beethoven X Optimism can deploy and added gitignore
